Iteration 9. Correction. Logger to middleware.
