+++ internal/logger/logger.go
@@ -1,73 +1,15 @@
 package logger
 
 import (
-	"net/http"
-	"time"
-
 	"go.uber.org/zap"
 )
 
-type (
-	responseData struct {
-		status int
-		size   int
-	}
-
-	loggingResponseWriter struct {
-		http.ResponseWriter
-		responseData *responseData
-	}
-)
-
-var Log *zap.SugaredLogger
-
-func Initialize() error {
+func NewLogger() *zap.SugaredLogger {
 	logger, err := zap.NewDevelopment()
 	if err != nil {
 		panic(err)
 	}
 	defer logger.Sync()
 
-	Log = logger.Sugar()
-	return nil
-}
-
-func RequestResponseLogger(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		start := time.Now()
-
-		responseData := &responseData{
-			status: 0,
-			size:   0,
-		}
-
-		lw := loggingResponseWriter{
-			ResponseWriter: w,
-			responseData:   responseData,
-		}
-
-		uri := r.RequestURI
-		method := r.Method
-		next.ServeHTTP(&lw, r)
-		duration := time.Since(start)
-		Log.Infow(
-			"Incoming request",
-			"uri", uri,
-			"method", method,
-			"status", responseData.status,
-			"duration", duration,
-			"size", responseData.size,
-		)
-	})
-}
-
-func (r *loggingResponseWriter) Write(b []byte) (int, error) {
-	size, err := r.ResponseWriter.Write(b)
-	r.responseData.size += size
-	return size, err
-}
-
-func (r *loggingResponseWriter) WriteHeader(statusCode int) {
-	r.ResponseWriter.WriteHeader(statusCode)
-	r.responseData.status = statusCode
+	return logger.Sugar()
 }
